rename(grades): 성적 관련 기능 폴더명 수정
